Split code and added RBS
